Procedural Spawning system upgraded to use seperate world and EntityManager.MoveEntitiesFrom

+++ UnstablePrototypes/BlobificationPrototype/Assets/ProceduralInstancing/ProceduralSpawnSystem.cs
@@ -1,4 +1,4 @@
-﻿using System.Linq;
+﻿using System;
 using Unity.Collections;
 using Unity.Jobs;
 using Unity.Mathematics;
@@ -19,14 +19,21 @@ public class ProceduralSpawnSystem : JobComponentSystem
 
     const float                        GridSize = 1.0F;
     const float                        PlantDensity = 0.5F;
-    const int                          MaxCreateChunksPerFrame = 4;
-    const int                          MaxDestroyChunksPerFrame = 8;
+    const int                          MaxCreateChunksPerFrame = 8;
+    const int                          MaxDestroyChunksPerFrame = 12;
+    const bool                         Deterministic = false;
 
-    struct GridChunk
+    struct GridChunk : IEquatable<int2>
     {
         public int2             Pos;
         public ComponentType    ChunkSceneType;
+
+        public bool Equals(int2 pos)
+        {
+            return Pos.Equals(pos);
+        }
     }
+
 
     //@TODO: This would be better expressed with a hashmap,
     //       however we currently can't iterate over all elements...
@@ -52,6 +59,10 @@ public class ProceduralSpawnSystem : JobComponentSystem
     [InjectComponentGroup] CollisionMeshInstances m_CollisionMeshes;
 
 
+    World m_ConstructionWorld;
+    EntityManager m_ConstructionManager;
+
+
     protected override void OnCreateManager(int capacity)
     {
         m_CreatedChunks = new NativeList<GridChunk>(capacity, Allocator.Persistent);
@@ -63,12 +74,18 @@ public class ProceduralSpawnSystem : JobComponentSystem
         prefab = AssetDatabase.LoadAssetAtPath<GameObject>("Assets/VegetationAssets/ECSVegetation0.prefab");
 
 
+
+        //@TODO: how does leak detection work for world creation?
+        m_ConstructionWorld = new World("Procedural construction world");
+        m_ConstructionManager = m_ConstructionWorld.GetOrCreateManager<EntityManager>();
         //@TODO: Would be nice if this wasn't necessary and transactions could increase capacity...
-        EntityManager.EntityCapacity = 1000 * 1000;
+        m_ConstructionManager.EntityCapacity = 1000 * 1000;
     }
 
     protected override void OnDestroyManager()
     {
+        m_ConstructionWorld.Dispose();
+
         m_CreatedChunks.Dispose();
         for (int i = 0;i != m_SpawnLocationCaches.Length;i++)
             m_SpawnLocationCaches[i].Dispose();
@@ -101,7 +118,7 @@ public class ProceduralSpawnSystem : JobComponentSystem
         }
     }
 
-    public int GetToBeCreatedGridPositions(NativeArray<int2> visible, NativeArray<int2> outGridPositions)
+    static int GetToBeCreatedGridPositions(NativeArray<GridChunk> createdChunks, NativeArray<int2> visible, NativeArray<int2> outGridPositions)
     {
         //@TODO: Foreach support
         // foreach (var gridPos in visible)
@@ -112,39 +129,22 @@ public class ProceduralSpawnSystem : JobComponentSystem
             //@TODO: Need Contains method here...
 
             ComponentType type;
-            if (!HasChunk(gridPos))
+            if (!createdChunks.Contains(gridPos))
                 outGridPositions[count++] = gridPos;
         }
 
         return count;
     }
 
-    public int GetToBeDestroyedGridPositions(NativeArray<int2> visible, NativeArray<int2> toBeDestroyed)
+    static int GetToBeDestroyedGridPositions(NativeArray<GridChunk> createdChunks, NativeArray<int2> visible, NativeArray<int2> toBeDestroyed)
     {
         int count = 0;
-        for (int i = 0; i != m_CreatedChunks.Length && count != toBeDestroyed.Length; i++)
+        for (int i = 0; i != createdChunks.Length && count != toBeDestroyed.Length; i++)
         {
-            if (!visible.Contains(m_CreatedChunks[i].Pos))
-                toBeDestroyed[count++] = m_CreatedChunks[i].Pos;
+            if (!visible.Contains(createdChunks[i].Pos))
+                toBeDestroyed[count++] = createdChunks[i].Pos;
         }
         return count;
-    }
-
-    public bool HasChunk(int2 gridPos)
-    {
-        return GetCreatedChunksIndex(gridPos) != -1;
-    }
-
-    public int GetCreatedChunksIndex(int2 gridPos)
-    {
-        for (int i = 0; i != m_CreatedChunks.Length; i++)
-        {
-            //@TODO: Thats a bit annoying...
-            if (math.all(m_CreatedChunks[i].Pos == gridPos))
-                return i;
-        }
-
-        return -1;
     }
 
     struct SpawnData
@@ -152,12 +152,39 @@ public class ProceduralSpawnSystem : JobComponentSystem
         public float3 Position;
     }
 
+    struct DoubleBufferCollisionInstancesJob : IJobParallelFor
+    {
+        //@TODO: How do we treat refcounts of the blob asset reference here????
+        [ReadOnly]
+        public ComponentDataArray<CollisionMeshInstance> SrcCollisionInstances;
+        public NativeArray<CollisionMeshInstance>        DstCollisionInstances;
+
+        public void Execute(int index)
+        {
+            DstCollisionInstances[index] = SrcCollisionInstances[index];
+            //@TODO: Doing this manually seems a bit annoying...
+            DstCollisionInstances[index].CollisionMesh.Retain();
+        }
+    }
+
+    struct DeallocateCollisionInstancesJob : IJobParallelFor
+    {
+        [DeallocateOnJobCompletion]
+        public NativeArray<CollisionMeshInstance>        CollisionInstances;
+
+        public void Execute(int index)
+        {
+            //@TODO: Doing this manually seems a bit annoying...
+            CollisionInstances[index].CollisionMesh.Release();
+        }
+    }
+
     //[ComputeJobOptimization]
     struct CalculateChunkSpawnLocationsJob : IJob
     {
         public int2                 ChunkPosition;
         [ReadOnly]
-        public ComponentDataArray<CollisionMeshInstance> CollisionInstances;
+        public NativeArray<CollisionMeshInstance> CollisionInstances;
 
         public NativeList<SpawnData>        SpawnLocations;
 
@@ -185,6 +212,35 @@ public class ProceduralSpawnSystem : JobComponentSystem
     }
 
 
+    //[ComputeJobOptimization]
+    struct CalculateToCreateAndDestroyChunks : IJob
+    {
+        //@TODO: Would be nice if this could be a local variable in the job...
+        public NativeList<int2> visibleGridPositions;
+
+        public NativeArray<GridChunk>  createdChunks;
+
+        public NativeList<int2> destroyGridPositions;
+        public NativeList<int2> createGridPositions;
+
+        public float2 CameraPosition;
+        public float  MaxDistance;
+        public float  GridSize;
+
+        public void Execute()
+        {
+            CalculateVisibleGridPositions(CameraPosition, MaxDistance, GridSize, visibleGridPositions);
+
+            // Destroys any invisible grid positions
+            destroyGridPositions.ResizeUninitialized(destroyGridPositions.Capacity);
+            destroyGridPositions.ResizeUninitialized(GetToBeDestroyedGridPositions(createdChunks, visibleGridPositions, destroyGridPositions));
+
+            createGridPositions.ResizeUninitialized(createGridPositions.Capacity);
+            createGridPositions.ResizeUninitialized(GetToBeCreatedGridPositions(createdChunks, visibleGridPositions, createGridPositions));
+        }
+    }
+
+
     struct PopulateChunk : IJob
     {
         //@TODO: This is just a hack. No data from the prefab gets transported except for ISharedComponentData
@@ -197,6 +253,7 @@ public class ProceduralSpawnSystem : JobComponentSystem
         //@TODO: Not supported. This is super annoying...
         //[DeallocateOnJobCompletion]
         public NativeList<SpawnData> SpawnLocations;
+
 
         public void Execute()
         {
@@ -214,25 +271,26 @@ public class ProceduralSpawnSystem : JobComponentSystem
         }
     }
 
-    void Destroy(int2 position)
+    void DestroyChunk(int2 position)
     {
         var chunkScene = new ProceduralChunkScene();
         chunkScene.Position = position;
 
-        Profiler.BeginSample("Destroy.CreateComponentGroup");
+        Profiler.BeginSample("DestroyChunk.CreateComponentGroup");
         var chunkSceneType = EntityManager.CreateSharedComponentType(chunkScene);
         var group = EntityManager.CreateComponentGroup(chunkSceneType);
         Profiler.EndSample();
 
         //@TODO: This is highly inconvenient...
-        Profiler.BeginSample("Destroy.GetEntities");
+        Profiler.BeginSample("DestroyChunk.GetEntities");
         var entityGroupArray = group.GetEntityArray();
         var entityArray = new NativeArray<Entity>(entityGroupArray.Length, Allocator.Temp, NativeArrayOptions.UninitializedMemory);
         entityGroupArray.CopyTo(entityArray);
         Profiler.EndSample();
 
-        Profiler.BeginSample("Destroy.DestroyEntity");
-        EntityManager.DestroyEntity(entityArray);
+        Profiler.BeginSample("DestroyChunk.DestroyEntity");
+        if (entityArray.Length != 0)
+            EntityManager.DestroyEntity(entityArray);
         Profiler.EndSample();
 
         entityArray.Dispose();
@@ -240,24 +298,24 @@ public class ProceduralSpawnSystem : JobComponentSystem
         group.Dispose();
 
         //@TODO: Need value based search function...
-        m_CreatedChunks.RemoveAtSwapBack(GetCreatedChunksIndex(position));
+        m_CreatedChunks.RemoveAtSwapBack(m_CreatedChunks.IndexOf(position));
     }
 
     //@TODO: Wow this is just insane amounts of code necessary for the simplest possible thing...
-    EntityArchetype GetArchetype(int2 chunkPosition, out ComponentType chunkSceneType)
+    void GetArchetype(int2 chunkPosition, out EntityArchetype archetype, out ComponentType chunkSceneType)
     {
         ComponentType[] types;
         Component[] components;
-        GameObjectEntity.GetComponents(EntityManager, prefab, false, out types, out components);
+        GameObjectEntity.GetComponents(m_ConstructionManager, prefab, false, out types, out components);
 
         ComponentType[] typesWithChunk = new ComponentType[types.Length + 1];
         types.CopyTo(typesWithChunk, 0);
 
         var chunkSceneComponent = new ProceduralChunkScene() { Position = chunkPosition };
-        chunkSceneType = EntityManager.CreateSharedComponentType(chunkSceneComponent);
+        chunkSceneType = m_ConstructionManager.CreateSharedComponentType(chunkSceneComponent);
         typesWithChunk[types.Length ] = chunkSceneType;
 
-        return EntityManager.CreateArchetype(typesWithChunk);
+        archetype = m_ConstructionManager.CreateArchetype(typesWithChunk);
     }
 
 
@@ -265,96 +323,130 @@ public class ProceduralSpawnSystem : JobComponentSystem
     {
         //@TODO: When is a good time for this to be called?
         Profiler.BeginSample("CommitTransaction");
-        EntityManager.CommitTransaction();
+
+        bool scheduleCreation = Deterministic || m_ConstructionManager.EntityTransactionDependency.IsCompleted;
+        if (scheduleCreation)
+            EntityManager.MoveEntitiesFrom(m_ConstructionManager);
         Profiler.EndSample();
 
         Profiler.BeginSample("CalculateVisible");
+        var calculateVisibleJob = new CalculateToCreateAndDestroyChunks();
+        calculateVisibleJob.createdChunks = m_CreatedChunks;
+        calculateVisibleJob.visibleGridPositions = new NativeList<int2>(512, Allocator.Temp);
+
         // Calculate all grid positions that should be visible
-        var visibleGridPositions = new NativeList<int2>(0, Allocator.Temp);
+        calculateVisibleJob.createGridPositions = new NativeList<int2>(MaxCreateChunksPerFrame, Allocator.Temp);
+        calculateVisibleJob.destroyGridPositions = new NativeList<int2>(MaxDestroyChunksPerFrame, Allocator.Temp);
         {
             var cameras = GetEntities<CameraEntity>();
+            //@TODO: What if camera doesnt exist
             if (cameras.Length != 0)
             {
                 float3 cameraPos = cameras[0].Transform.position;
-                float cullingDistance = cameras[0].View.Distance;
-                CalculateVisibleGridPositions(cameraPos.xz, cullingDistance, GridSize, visibleGridPositions);
+                calculateVisibleJob.CameraPosition = cameraPos.xz;
+                calculateVisibleJob.GridSize = GridSize;
+                calculateVisibleJob.MaxDistance = cameras[0].View.Distance;
             }
         }
         Profiler.EndSample();
 
+
+        calculateVisibleJob.Run();
+
         Profiler.BeginSample("DestroyChunk");
-        // Destroys any invisible grid positions
-        var destroyGridPositions = new NativeArray<int2>(MaxDestroyChunksPerFrame, Allocator.Temp);
-        
-        Profiler.BeginSample("Destroy.FindToBeDestroyed");
-        int toDestroyedCount = GetToBeDestroyedGridPositions(visibleGridPositions, destroyGridPositions);
-        Profiler.EndSample();
-        
-        for (int i = 0; i != toDestroyedCount; i++)
-            Destroy(destroyGridPositions[i]);
-        Profiler.EndSample();
-
-
-        Profiler.BeginSample("ScheduleChunkCreation");
-        var toBeCreatedChunks = new NativeArray<int2>(MaxCreateChunksPerFrame, Allocator.Temp);
-        Profiler.BeginSample("FindToBeCreated");
-        int toBeCreatedCount = GetToBeCreatedGridPositions(visibleGridPositions, toBeCreatedChunks);
+        {
+            for (int i = 0; i != calculateVisibleJob.destroyGridPositions.Length; i++)
+                DestroyChunk(calculateVisibleJob.destroyGridPositions[i]);
+        }
         Profiler.EndSample();
 
         // Schedule jobs for each grid chunk that became visible
-
-        if (toBeCreatedCount != 0)
+        if (scheduleCreation)
         {
-            // Extract all archetypes first, because extracting the archetype will CommitTransaction...
-            // @TODO: When Simons SharedComponentData refactor is ready this should be unnecessary...
-            var archetypes = new NativeArray<EntityArchetype>(toBeCreatedCount, Allocator.Temp);
-            var chunkSceneTypes = new NativeArray<ComponentType>(toBeCreatedCount, Allocator.Temp);
-            for (int i = 0; i != toBeCreatedCount; i++)
+            if (calculateVisibleJob.createGridPositions.Length != 0)
             {
-                ComponentType type;
-                archetypes[i] = GetArchetype(toBeCreatedChunks[i], out type);
-                chunkSceneTypes[i] = type;
+                UpdateInjectedComponentGroups();
+
+                Profiler.BeginSample("CreateChunks");
+                dependency = ScheduleCreateChunks(dependency, calculateVisibleJob.createGridPositions);
+                Profiler.EndSample();
             }
 
-            UpdateInjectedComponentGroups();
-
-            var spawnInputDependency = dependency;
-            dependency = JobHandle.CombineDependencies(EntityManager.EntityTransactionDependency, dependency);
-
-            var transaction = EntityManager.BeginTransaction();
-            for (int i=0;i != toBeCreatedCount;i++)
-            {
-                var createPos = toBeCreatedChunks[i];
-
-                var spawnLocationJob = new CalculateChunkSpawnLocationsJob();
-                spawnLocationJob.ChunkPosition = createPos;
-                spawnLocationJob.CollisionInstances = m_CollisionMeshes.Instances;
-                spawnLocationJob.SpawnLocations = m_SpawnLocationCaches[i];
-                spawnLocationJob.SpawnLocations.Clear();
-
-                var spawnDependency = spawnLocationJob.Schedule(spawnInputDependency);
-
-                var chunkJob = new PopulateChunk();
-                chunkJob.Archetype = archetypes[i];
-                chunkJob.EntityTransaction = transaction;
-                chunkJob.ChunkPosition = createPos;
-                chunkJob.SpawnLocations = spawnLocationJob.SpawnLocations;
-
-                m_CreatedChunks.Add(new GridChunk(){ Pos = createPos, ChunkSceneType = chunkSceneTypes[i] });
-
-                dependency = chunkJob.Schedule(JobHandle.CombineDependencies(spawnDependency, dependency));
-            }
-
-            EntityManager.EntityTransactionDependency = dependency;
-
-            archetypes.Dispose();
-            chunkSceneTypes.Dispose();
+            //@TODO: Error message produced from this seems wrong...
+            //toBeCreatedChunks.Dispose();
         }
-        toBeCreatedChunks.Dispose();
-        Profiler.EndSample();
 
-        destroyGridPositions.Dispose();
-        visibleGridPositions.Dispose();
+        calculateVisibleJob.createGridPositions.Dispose();
+        calculateVisibleJob.destroyGridPositions.Dispose();
+        calculateVisibleJob.visibleGridPositions.Dispose();
+
+        return dependency;
+    }
+
+    JobHandle ScheduleCreateChunks(JobHandle dependency, NativeArray<int2> toBeCreatedChunks)
+    {
+        // Double buffer collision instances so that this job can run truly async
+        var doubleBufferCollision = new DoubleBufferCollisionInstancesJob();
+        doubleBufferCollision.DstCollisionInstances = new NativeArray<CollisionMeshInstance>(m_CollisionMeshes.Instances.Length, Allocator.Persistent, NativeArrayOptions.UninitializedMemory);
+        doubleBufferCollision.SrcCollisionInstances = m_CollisionMeshes.Instances;
+        dependency = doubleBufferCollision.Schedule(m_CollisionMeshes.Instances.Length, 16, dependency);
+
+        // Extract all archetypes first, because extracting the archetype will CommitTransaction...
+        // @TODO: When Simons SharedComponentData refactor is ready this should be unnecessary...
+        var archetypes = new NativeArray<EntityArchetype>(toBeCreatedChunks.Length, Allocator.Temp);
+        var chunkSceneTypes = new NativeArray<ComponentType>(toBeCreatedChunks.Length, Allocator.Temp);
+        for (int i = 0; i != toBeCreatedChunks.Length; i++)
+        {
+            ComponentType type;
+            EntityArchetype arch;
+            GetArchetype(toBeCreatedChunks[i], out arch, out type);
+            chunkSceneTypes[i] = type;
+            archetypes[i] = arch;
+        }
+
+        var spawnDependencies = new NativeArray<JobHandle>(toBeCreatedChunks.Length, Allocator.Temp);
+
+        var transaction = m_ConstructionManager.BeginTransaction();
+        for (int i = 0; i != toBeCreatedChunks.Length; i++)
+        {
+            var createPos = toBeCreatedChunks[i];
+
+            var spawnLocationJob = new CalculateChunkSpawnLocationsJob();
+            spawnLocationJob.ChunkPosition = createPos;
+            spawnLocationJob.CollisionInstances = doubleBufferCollision.DstCollisionInstances;
+            spawnLocationJob.SpawnLocations = m_SpawnLocationCaches[i];
+            spawnLocationJob.SpawnLocations.Clear();
+
+            spawnDependencies[i] = spawnLocationJob.Schedule(dependency);
+        }
+
+        for (int i = 0; i != toBeCreatedChunks.Length; i++)
+        {
+            var createPos = toBeCreatedChunks[i];
+
+            GridChunk gridChunk;
+            gridChunk.Pos = createPos;
+            gridChunk.ChunkSceneType = chunkSceneTypes[i];
+
+            PopulateChunk chunkJob;
+            chunkJob.EntityTransaction = transaction;
+            chunkJob.ChunkPosition = createPos;
+            chunkJob.SpawnLocations = m_SpawnLocationCaches[i];
+            chunkJob.Archetype = archetypes[i];
+
+            m_CreatedChunks.Add(gridChunk);
+
+            var chunkDependency = JobHandle.CombineDependencies(spawnDependencies[i], m_ConstructionManager.EntityTransactionDependency);
+            m_ConstructionManager.EntityTransactionDependency = chunkJob.Schedule(chunkDependency);
+        }
+
+        spawnDependencies.Dispose();
+        archetypes.Dispose();
+        chunkSceneTypes.Dispose();
+
+        var deallocateCollisionJob = new DeallocateCollisionInstancesJob();
+        deallocateCollisionJob.CollisionInstances = doubleBufferCollision.DstCollisionInstances;
+        m_ConstructionManager.EntityTransactionDependency = deallocateCollisionJob.Schedule(deallocateCollisionJob.CollisionInstances.Length, 16, m_ConstructionManager.EntityTransactionDependency);
 
         return dependency;
     }
